Cherry-picked from d40f179c1dab299dc6540d7c73a1ba7b398d976d: Add redeploy scripts from Amazon test environments

+++ test/system-test/src/test/scripts/redeploy-axel.sh
@@ -18,8 +18,22 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>
 #
 
+echo "Script kördes med följande parameter: $0 $@"
+
+export JAVA_MAX_MEM=1024m
+export JAVA_MAX_PERM_MEM=384m
+export JAVA_HOME=/usr/java/jdk1.7.0_45
 
 AXEL="axel-1.0-SNAPSHOT"
+AXEL_USER=axel
+
+if [ $# -eq 0 ]
+  then
+    echo "ERROR: Parameter för Mongo databasnamn saknas"
+    exit 1
+  else
+    MONGO_DB_NAME=$1
+fi
 
 BASEDIR=`dirname $0`
 BASEDIR="`cd \"$BASEDIR\" 2>/dev/null && pwd`"
@@ -30,25 +44,54 @@ cd $BASEDIR
 echo "pwd="
 pwd
 
+if [ -d $AXEL ]; then
+    if [ -x $AXEL/bin/stop ] && pgrep -u $AXEL_USER -f "java.*$AXEL" > /dev/null; then
+	echo "Stoppar axel"
+	$AXEL/bin/stop
+	echo "sleep 20"
+	sleep 20
+    fi
+    echo "Tar bort gamla axel"
+    rm -rf $AXEL
+fi
 
-echo "stoppar axel"
-$AXEL/bin/stop
+echo "Packar upp axel"
+tar xvfz $AXEL.tar.gz 
+if [ $? -ne 0 ]; then
+    echo "Kunde inte packa upp axel"
+    exit 1
+fi
 
-echo "sleep 20"
-sleep 20
+echo "Skapar Mongo indexer"
+$AXEL/docs/mongo/createIndexes.sh $MONGO_DB_NAME
+if [ $? -ne 0 ]; then
+    echo "Kunde inte skapar Mongo indexer"
+    exit 1
+fi
 
-echo "tar bort gamla axel"
-rm -rf $AXEL
+echo "Startar axel"
+nohup $AXEL/bin/start &
 
-echo "packar upp axel"
-tar xvfz $AXEL.tar.gz
+# Kill script if Axel has not started in 15 minutes
+sleep $((15*60)) && kill $$ &
+watchdogpid=$!
 
+while true
+do
+        status_code=$(curl --write-out %{http_code} --silent --output "healthList.json" -u 'admin:admin' -X POST -d '{"type":"exec", "mbean":"se.inera.axel:name=axel,service=Health,type=HealthView", "operation":"healthList()" }' -H "Accept: application/json" -i http://localhost:8181/hawtio/jolokia)
 
-echo "startar axel"
-$AXEL/bin/start
+        if [ $status_code -eq 200 ]; then
+                if grep -iq 'ERROR' healthList.json; then
+                        echo "Health view reported errors"
+                else
+                        echo "No errors reported Axel is started"
+                        sleep 5
+                        break
+                fi
+        fi
+        echo "Axel is not started sleeping 10 seconds"
+        sleep 10
+done
 
-echo "väntar ett bra tag"
-sleep 60
-
-echo "exiterar"
-
+kill $watchdogpid
+echo "Exiterar"
